requirements.txt
PyQt5
forex-python
pillow
requests
lxml
bs4
pyowm
geopy